NIFI-9798 Remove with configuration to pubsub lite processors and remove duplicate properties
