+++ nifi-nar-bundles/nifi-gcp-bundle/nifi-gcp-processors/src/main/java/org/apache/nifi/processors/gcp/pubsub/AbstractGCPubSubProcessor.java
@@ -16,16 +16,9 @@
  */
 package org.apache.nifi.processors.gcp.pubsub;
 
-import com.google.api.core.ApiFunction;
-import com.google.api.gax.rpc.TransportChannelProvider;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.cloud.ServiceOptions;
 
-import com.google.cloud.pubsub.v1.TopicAdminSettings;
-import io.grpc.HttpConnectProxiedSocketAddress;
-import io.grpc.ManagedChannelBuilder;
-import io.grpc.ProxiedSocketAddress;
-import io.grpc.ProxyDetector;
 import org.apache.nifi.components.PropertyDescriptor;
 import org.apache.nifi.components.ValidationContext;
 import org.apache.nifi.components.ValidationResult;
@@ -34,19 +27,11 @@ import org.apache.nifi.processor.Relationship;
 import org.apache.nifi.processor.VerifiableProcessor;
 import org.apache.nifi.processor.util.StandardValidators;
 import org.apache.nifi.processors.gcp.AbstractGCPProcessor;
-import org.apache.nifi.processors.gcp.ProxyAwareTransportFactory;
-import org.apache.nifi.proxy.ProxyConfiguration;
 
-import javax.annotation.Nullable;
-import java.io.IOException;
-import java.net.InetSocketAddress;
-import java.net.Proxy;
-import java.net.SocketAddress;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 public abstract class AbstractGCPubSubProcessor extends AbstractGCPProcessor implements VerifiableProcessor {
@@ -74,19 +59,6 @@ public abstract class AbstractGCPubSubProcessor extends AbstractGCPProcessor imp
     private static final Set<Relationship> relationships = Collections.unmodifiableSet(new HashSet<>(Arrays.asList(REL_SUCCESS, REL_FAILURE)));
 
     @Override
-    public List<PropertyDescriptor> getSupportedPropertyDescriptors() {
-        return Collections.unmodifiableList(Arrays.asList(
-                PROJECT_ID,
-                GCP_CREDENTIALS_PROVIDER_SERVICE,
-                PROXY_HOST,
-                PROXY_PORT,
-                HTTP_PROXY_USERNAME,
-                HTTP_PROXY_PASSWORD,
-                ProxyConfiguration.createProxyConfigPropertyDescriptor(true, ProxyAwareTransportFactory.PROXY_SPECS))
-        );
-    }
-
-    @Override
     public Set<Relationship> getRelationships() {
         return relationships;
     }
@@ -112,49 +84,4 @@ public abstract class AbstractGCPubSubProcessor extends AbstractGCPProcessor imp
         return results;
     }
 
-    protected TransportChannelProvider getTransportChannelProvider(ProcessContext context) {
-        final ProxyConfiguration proxyConfiguration = ProxyConfiguration.getConfiguration(context, () -> {
-            final String proxyHost = context.getProperty(PROXY_HOST).evaluateAttributeExpressions().getValue();
-            final Integer proxyPort = context.getProperty(PROXY_PORT).evaluateAttributeExpressions().asInteger();
-            if (proxyHost != null && proxyPort != null && proxyPort > 0) {
-                final ProxyConfiguration componentProxyConfig = new ProxyConfiguration();
-                final String proxyUser = context.getProperty(HTTP_PROXY_USERNAME).evaluateAttributeExpressions().getValue();
-                final String proxyPassword = context.getProperty(HTTP_PROXY_PASSWORD).evaluateAttributeExpressions().getValue();
-                componentProxyConfig.setProxyType(Proxy.Type.HTTP);
-                componentProxyConfig.setProxyServerHost(proxyHost);
-                componentProxyConfig.setProxyServerPort(proxyPort);
-                componentProxyConfig.setProxyUserName(proxyUser);
-                componentProxyConfig.setProxyUserPassword(proxyPassword);
-                return componentProxyConfig;
-            }
-            return ProxyConfiguration.DIRECT_CONFIGURATION;
-        });
-
-        return TopicAdminSettings.defaultGrpcTransportProviderBuilder()
-                .setChannelConfigurator(new ApiFunction<ManagedChannelBuilder, ManagedChannelBuilder>() {
-                    @Override
-                    public ManagedChannelBuilder apply(ManagedChannelBuilder managedChannelBuilder) {
-                        return managedChannelBuilder.proxyDetector(
-                                new ProxyDetector() {
-                                    @Nullable
-                                    @Override
-                                    public ProxiedSocketAddress proxyFor(SocketAddress socketAddress)
-                                            throws IOException {
-                                        if (Proxy.Type.HTTP.equals(proxyConfiguration.getProxyType())) {
-                                            return HttpConnectProxiedSocketAddress.newBuilder()
-                                                    .setUsername(proxyConfiguration.getProxyUserName())
-                                                    .setPassword(proxyConfiguration.getProxyUserPassword())
-                                                    .setProxyAddress(new InetSocketAddress(proxyConfiguration.getProxyServerHost(),
-                                                            proxyConfiguration.getProxyServerPort()))
-                                                    .setTargetAddress((InetSocketAddress) socketAddress)
-                                                    .build();
-                                        } else {
-                                            return null;
-                                        }
-                                    }
-                                });
-                    }
-                })
-                .build();
-    }
 }
